requirements.txt
django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
python-dotenv==1.1.0
transformers==4.51.3
torch==2.7.0
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.59
openai==1.78.0
numpy==2.2.5
regex==2024.11.6
tqdm==4.67.1 